+++ Cavestory/Globals.h
@@ -14,9 +14,9 @@ enum Direction {
 };
 
 struct Vector2 {
-	int x, y;
+	float x, y;
 	Vector2() : x(0), y(0) {}
-	Vector2(int x, int y) : x(x), y(y) {}
+	Vector2(float x, float y) : x(x), y(y) {}
 	Vector2 zero() {
 		return Vector2(0, 0);
 	}
