+++ Cavestory/Sprite.cpp
@@ -7,7 +7,7 @@ Sprite::Sprite()
 {
 }
 
-Sprite::Sprite(Graphics & graphics, const std::string & filePath, int sourceX, int sourceY, int width, int height, int posX, int posY) :
+Sprite::Sprite(Graphics & graphics, const std::string & filePath, int sourceX, int sourceY, int width, int height, float posX, float posY) :
 	_x(posX),
 	_y(posY)
 {
